Move compiler-based resolver tests to compiler
